Redid population of door triggers

+++ main/source/detour/DetourTileCacheBuilder.cpp
@@ -1775,12 +1775,12 @@ dtStatus dtBuildTileCachePolyMesh(dtTileCacheAlloc* alloc,
 	if (!mesh.areas)
 		return DT_FAILURE | DT_OUT_OF_MEMORY;
 
-	mesh.flags = (unsigned short*)alloc->alloc(sizeof(unsigned short)*maxTris);
+	mesh.flags = (unsigned int*)alloc->alloc(sizeof(unsigned int)*maxTris);
 	if (!mesh.flags)
 		return DT_FAILURE | DT_OUT_OF_MEMORY;
 
 	// Just allocate and clean the mesh flags array. The user is resposible for filling it.
-	memset(mesh.flags, 0, sizeof(unsigned short) * maxTris);
+	memset(mesh.flags, 0, sizeof(unsigned int) * maxTris);
 		
 	mesh.nverts = 0;
 	mesh.npolys = 0;
